Move backoffice course views to dedicated module
